feat(web): make skill detail pages indexable via intercepting routes
Replace the query-param-driven modal (?skill=slug + pushState) with a
real /[slug] route plus a parallel @modal slot that intercepts it, so
search engines can crawl full skill content while the modal experience
stays unchanged.

- Add standalone /[slug] detail page with canonical/OG/Twitter metadata
  and JSON-LD; unknown slugs return 404
- Add @modal/(.)[slug] intercepting route that renders the existing modal
  over the homepage on client navigation
- Extract SkillMarkdown, SkillMetaPills, and SkillInstallAction so the
  page and modal share one rendering
- Move stats into a shared StatsProvider fed from the root layout
- List every /[slug] in sitemap.xml and redirect old ?skill= links
- Preserve scroll position when opening a card (router.push scroll:false)
