+++ web/src/app/page.tsx
@@ -1,37 +1,19 @@
-import { getSkillBySlug, getSkills } from '@/lib/skills';
-import { getSkillStats } from '@/lib/skill-stats';
+import { redirect } from 'next/navigation';
+import { getSkills } from '@/lib/skills';
 import { Marketplace } from '@/components/Marketplace';
 import { Suspense } from 'react';
-import type { Metadata } from 'next';
 
 interface HomeProps {
   searchParams: Promise<{ skill?: string }>;
 }
 
-export async function generateMetadata({ searchParams }: HomeProps): Promise<Metadata> {
+export default async function Home({ searchParams }: HomeProps) {
+  const skills = await getSkills();
   const { skill } = await searchParams;
 
-  if (!skill) {
-    return {};
+  if (skill) {
+    redirect(`/${skill}`);
   }
-
-  const entry = await getSkillBySlug(skill);
-
-  if (!entry) {
-    return {};
-  }
-
-  const name = entry.metadata?.name ?? entry.name;
-
-  return {
-    title: `${name} · Flc's Skills`,
-    description: entry.metadata?.description ?? entry.description,
-  };
-}
-
-export default async function Home() {
-  const skills = await getSkills();
-  const initialStats = await getSkillStats(skills.map((skill) => skill.slug));
 
   return (
     <div className="min-h-screen">
@@ -49,7 +31,7 @@ export default async function Home() {
           </div>
         </div>
       }>
-        <Marketplace initialSkills={skills} initialStats={initialStats} />
+        <Marketplace initialSkills={skills} />
       </Suspense>
     </div>
   );
