+++ web/src/app/sitemap.ts
@@ -1,12 +1,24 @@
 import type { MetadataRoute } from 'next';
+import { getSkills } from '@/lib/skills';
+import { parseSkillMetadataDate } from '@/lib/utils';
 
-export default function sitemap(): MetadataRoute.Sitemap {
+const BASE_URL = 'https://skills.flc.io';
+
+export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
+  const skills = await getSkills();
+
   return [
     {
-      url: 'https://skills.flc.io',
+      url: BASE_URL,
       lastModified: new Date(),
       changeFrequency: 'weekly',
       priority: 1,
     },
+    ...skills.map((skill) => ({
+      url: `${BASE_URL}/${skill.slug}`,
+      lastModified: parseSkillMetadataDate(skill.metadata?.created) ?? new Date(),
+      changeFrequency: 'weekly' as const,
+      priority: 0.8,
+    })),
   ];
 }
